backend/requirements.txt
fastapi
uvicorn
sqlalchemy
dotenv
pymysql
alembic
cryptography
pydantic[email]
passlib
python-jose
fastapi-security
argon2-cffi
python-multipart